+++ Libplanet/Tx/Transaction.cs
@@ -25,7 +25,7 @@ namespace Libplanet.Tx
     /// </typeparam>
     /// <seealso cref="IAction"/>
     /// <seealso cref="PolymorphicAction{T}"/>
-    public sealed class Transaction<T> : IEquatable<Transaction<T>>, ITxMetadata
+    public sealed class Transaction<T> : IEquatable<Transaction<T>>, ITxExcerpt
         where T : IAction, new()
     {
         private const string TimestampFormat = "yyyy-MM-ddTHH:mm:ss.ffffffZ";